Attempt to move MultivariateIntPolynomial to its own file, does not compile
