+++ symengine/CMakeLists.txt
@@ -43,6 +43,7 @@ set(SRC
     parser.cpp
     mp_wrapper.cpp
     sets.cpp
+    polynomial_multivariate.cpp
 )
 
 if (WITH_MPFR)
@@ -67,7 +68,7 @@ endif()
 
 # Needed for "make install"
 set(HEADERS
-    add.h        symengine_assert.h  functions.h  monomials.h  polynomial.h  number.h    rings.h
+    add.h        symengine_assert.h  functions.h  monomials.h  polynomial.h  polynomial_multivariate.h  number.h    rings.h
     basic.h      symengine_rcp.h     integer.h    mul.h        pow.h       constants.h
     symbol.h     expression.h        parser.h     mp_class.h   mp_wrapper.h
     basic-inl.h  dict.h           matrix.h     ntheory.h    rational.h complex.h
